requirements.txt
sounddevice==0.5.5
playsound==1.3.0
numpy==2.4.2
PyObjC==12.1